+++ fractal/components/drawers/Drawer/Drawer.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+    "title": "Drawer",
+    "label": "Drawer",
+    "status": "ready",
+    "collated": true,
+    "preview": "@drawer-preview",
+    "context": {
+        name: 'c-site-drawer',
+        trigger: 'Lorem Ipsum'
+    }
+};