+++ fractal/components/drawers/DrawerPane/DrawerPane.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    "title": "Drawer Pane",
+    "label": "Drawer Pane",
+    "status": "ready",
+    "collated": true,
+    "preview": "@drawer-preview",
+    "context": {
+        children: 'Lorem Ipsum'
+    }
+};