+++ fractal/components/drawers/SiteDrawer/SiteDrawer.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    "title": "Site Drawer",
+    "label": "Site Drawer",
+    "status": "ready",
+    "collated": true,
+    "preview": "@drawer-preview",
+    "context": {
+        trigger: "Lorem Ipsum"
+    }
+};