+++ fractal/components/navigation-bars/SiteNavigation/SiteNavigation.config.js
@@ -1,13 +1,17 @@
-const navigationConfig = require('../NavigationBar/NavigationBar.config');
-
 module.exports = {
     "title": "Site Navigation",
     "label": "Site Navigation",
     "order": 5,
     "status": "ready",
     "collated": true,
-    "preview": "@site-navigation-preview",
+    "preview": "@site-navigation-bar-preview",
     "context": {
-        tabs: navigationConfig.context.tabs
+        tabs: [
+            {label: "Find a Doctor", to: "http://www.somewhere.com"},
+            {label: "Health Experts", to: "http://www.elsewhere.com", selected: true},
+            {label: "Topics", to: "http://www.everywhere.com"},
+            {label: "Video", to: "http://www.nowhere.com"},
+            {label: "Health Tools", to: "http://www.where.com"}
+        ]
     }
 };