+++ fractal/components/navigation-bars/SiteNavigationBar/SiteNavigationBar.config.js
@@ -1,11 +1,13 @@
+const navigationConfig = require('../SiteNavigation/SiteNavigation.config');
+
 module.exports = {
     "title": "Site Navigation Bar",
     "label": "Site Navigation Bar",
     "order": 6,
     "status": "ready",
     "collated": true,
-    "preview": "@site-navigation-preview",
+    "preview": "@site-navigation-bar-preview",
     "context": {
-
+        tabs: navigationConfig.context.tabs
     }
 };