+++ fractal/components/site-logos/SiteLogo/SiteLogo.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    "title": "Site Logo",
+    "label": "Site Logo",
+    "status": "ready",
+    "collated": true,
+    "preview": "@site-navigation-bar-preview",
+    "context": {}
+};